+++ pkg/core/resources/apis/mesh/traffic_route.go
@@ -44,15 +44,15 @@ func (t *TrafficRouteResource) SetSpec(spec model.ResourceSpec) error {
 var _ model.ResourceList = &TrafficRouteResourceList{}
 
 type TrafficRouteResourceList struct {
-	Total      uint64
+	Total      uint32
 	Items      []*TrafficRouteResource
 	Pagination model.Pagination
 }
 
-func (l *TrafficRouteResourceList) GetTotal() uint64 {
+func (l *TrafficRouteResourceList) GetTotal() uint32 {
 	return l.Total
 }
-func (l *TrafficRouteResourceList) SetTotal(total uint64) {
+func (l *TrafficRouteResourceList) SetTotal(total uint32) {
 	l.Total = total
 }
 func (l *TrafficRouteResourceList) GetItems() []model.Resource {
